+++ regimes/it/scenarios.go
@@ -30,6 +30,13 @@ const (
 // involved, we've focussed here on the most useful.
 var invoiceTags = common.InvoiceTagsWith([]*cbc.KeyDefinition{
 	// *** Document Type Tags ***
+	// Sales:
+	{
+		Key: tax.TagB2G,
+		Name: i18n.String{
+			i18n.EN: "Business to Government",
+		},
+	},
 	{
 		Key: TagFreelance,
 		Name: i18n.String{
@@ -37,6 +44,7 @@ var invoiceTags = common.InvoiceTagsWith([]*cbc.KeyDefinition{
 			i18n.IT: "Parcella",
 		},
 	},
+	// Self-billed:
 	{
 		Key: TagCeilingExceeded,
 		Name: i18n.String{
@@ -114,12 +122,6 @@ var invoiceTags = common.InvoiceTagsWith([]*cbc.KeyDefinition{
 			i18n.IT: "Beni ammortizzabili",
 		},
 	},
-	{
-		Key: tax.TagB2G,
-		Name: i18n.String{
-			i18n.EN: "Business to Government",
-		},
-	},
 })
 
 var scenarios = []*tax.ScenarioSet{
@@ -131,7 +133,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 	List: []*tax.Scenario{
 		// **** DOCUMENT FORMAT ****
 		{
-			// default
 			Name: i18n.String{
 				i18n.EN: "Private Invoice",
 				i18n.IT: "Fattura Privata",
@@ -152,14 +153,10 @@ var invoiceScenarios = &tax.ScenarioSet{
 		},
 		// **** TIPO DOCUMENTO ****
 		{
-			// Default
 			Types: []cbc.Key{bill.InvoiceTypeStandard},
 			Name: i18n.String{
 				i18n.EN: "Regular Invoice",
 				i18n.IT: "Fattura",
-			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD01", // deprecation pending
 			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD01",
@@ -172,9 +169,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Advance or down payment on invoice",
 				i18n.IT: "Acconto / anticipo su fattura",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD02",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD02",
 			},
@@ -185,9 +179,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Credit Note",
 				i18n.IT: "Nota di credito",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD04",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD04",
 			},
@@ -197,9 +188,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 			Name: i18n.String{
 				i18n.EN: "Debit Note",
 				i18n.IT: "Nota di debito",
-			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD05",
 			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD05",
@@ -212,9 +200,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Freelancer invoice with retained taxes",
 				i18n.IT: "Parcella",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD06",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD06",
 			},
@@ -225,9 +210,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 			Name: i18n.String{
 				i18n.EN: "Advance or down payment on freelance invoice",
 				i18n.IT: "Acconto / anticipo su parcella",
-			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD03",
 			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD03",
@@ -240,9 +222,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Simplified Invoice",
 				i18n.IT: "Fattura Semplificata",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD07",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD07",
 			},
@@ -253,9 +232,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 			Name: i18n.String{
 				i18n.EN: "Simplified Credit Note",
 				i18n.IT: "Nota di credito semplificata",
-			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD08",
 			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD08",
@@ -268,9 +244,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Simplified Debit Note",
 				i18n.IT: "Nota di debito semplificata",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD09",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD09",
 			},
@@ -282,11 +255,8 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Self-billed for self consumption or for free transfer without recourse",
 				i18n.IT: "Fattura per autoconsumo o per cessioni gratuite senza rivalsa",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD27", // order is important
-			},
 			Ext: tax.Extensions{
-				ExtKeySDIDocumentType: "TD27",
+				ExtKeySDIDocumentType: "TD27", // order is important
 			},
 		},
 		{
@@ -295,9 +265,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 			Name: i18n.String{
 				i18n.EN: "Reverse charge",
 				i18n.IT: "Integrazione fattura reverse charge interno",
-			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD16",
 			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD16",
@@ -310,9 +277,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Self-billed Import",
 				i18n.IT: "Integrazione/autofattura per acquisto servizi da estero",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD17",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD17",
 			},
@@ -323,9 +287,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 			Name: i18n.String{
 				i18n.EN: "Self-billed EU Goods Import",
 				i18n.IT: "Integrazione per acquisto beni intracomunitari",
-			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD18",
 			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD18",
@@ -338,9 +299,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Self-billed Goods Import",
 				i18n.IT: "Integrazione/autofattura per acquisto beni ex art.17 c.2 DPR 633/72",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD19",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD19",
 			},
@@ -351,9 +309,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 			Name: i18n.String{
 				i18n.EN: "Self-billed Regularization",
 				i18n.IT: "Autofattura per regolarizzazione e integrazione delle fatture - art.6 c.8 d.lgs.471/97 o art.46 c.5 D.L.331/93",
-			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD20",
 			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD20",
@@ -366,9 +321,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Self-billed invoice when ceiling exceeded",
 				i18n.IT: "Autofattura per splafonamento",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD21",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD21",
 			},
@@ -379,9 +331,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 			Name: i18n.String{
 				i18n.EN: "Self-billed for goods extracted from VAT warehouse",
 				i18n.IT: "Estrazione beni da Deposito IVA",
-			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD22",
 			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD22",
@@ -394,9 +343,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Self-billed for goods extracted from VAT warehouse with VAT payment",
 				i18n.IT: "Estrazione beni da Deposito IVA con versamento IVA",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD23",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD23",
 			},
@@ -407,9 +353,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 			Name: i18n.String{
 				i18n.EN: "Deferred invoice ex art.21, c.4, lett. a) DPR 633/72",
 				i18n.IT: "Fattura differita - art.21 c.4 lett. a",
-			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD24",
 			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD24",
@@ -422,9 +365,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Deferred invoice ex art.21, c.4, third period lett. b) DPR 633/72",
 				i18n.IT: "Fattura differita - art.21 c.4 terzo periodo lett. b",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD25",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD25",
 			},
@@ -436,9 +376,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 				i18n.EN: "Sale of depreciable assets and for internal transfers (ex art.36 DPR 633/72",
 				i18n.IT: "Cessione di beni ammortizzabili e per passaggi interni - art.36 DPR 633/72",
 			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD26",
-			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD26",
 			},
@@ -449,9 +386,6 @@ var invoiceScenarios = &tax.ScenarioSet{
 			Name: i18n.String{
 				i18n.EN: "Purchases from San Marino with VAT (paper invoice)",
 				i18n.IT: "Acquisti da San Marino con IVA (fattura cartacea)",
-			},
-			Codes: cbc.CodeMap{
-				KeyFatturaPATipoDocumento: "TD28",
 			},
 			Ext: tax.Extensions{
 				ExtKeySDIDocumentType: "TD28",