+++ regimes/mx/mx.go
@@ -24,10 +24,9 @@ func init() {
 
 // Custom keys used typically in meta or codes information.
 const (
-	KeySATFormaPago         cbc.Key = "sat-forma-pago"          // for mapping to c_FormaPago’s codes
-	KeySATTipoDeComprobante cbc.Key = "sat-tipo-de-comprobante" // for mapping to c_TipoDeComprobante’s codes
-	KeySATTipoRelacion      cbc.Key = "sat-tipo-relacion"       // for mapping to c_TipoRelacion’s codes
-	KeySATImpuesto          cbc.Key = "sat-impuesto"            // for mapping to c_Impuesto’s codes
+	KeySATFormaPago    cbc.Key = "sat-forma-pago"    // for mapping to c_FormaPago’s codes
+	KeySATTipoRelacion cbc.Key = "sat-tipo-relacion" // for mapping to c_TipoRelacion’s codes
+	KeySATImpuesto     cbc.Key = "sat-impuesto"      // for mapping to c_Impuesto’s codes
 )
 
 // Official SAT and CFDI codes to include in stamps.