+++ tax/set.go
@@ -50,10 +50,6 @@ func (c *Combo) ValidateWithContext(ctx context.Context) error {
 			r.InCategories(),
 		),
 		validation.Field(&c.Rate,
-			validation.When(
-				(cat != nil && cat.RateRequired),
-				validation.Required,
-			),
 			r.InCategoryRates(c.Category),
 		),
 		validation.Field(&c.Ext,
@@ -75,6 +71,11 @@ func (c *Combo) ValidateWithContext(ctx context.Context) error {
 	)
 	if err != nil {
 		return err
+	}
+	if cat != nil && cat.Validation != nil {
+		if err := cat.Validation(c); err != nil {
+			return err
+		}
 	}
 	return r.ValidateObject(c)
 }
@@ -125,6 +126,17 @@ func (c *Combo) prepare(r *Regime, tags []cbc.Key, date cal.Date) error {
 	if rate == nil {
 		return ErrInvalidRate.WithMessage("'%s' rate not defined in category '%s'", c.Rate.String(), c.Category.String())
 	}
+
+	// Copy over the predefined extensions from the rate to the combo.
+	if len(rate.Ext) > 0 {
+		if c.Ext == nil {
+			c.Ext = make(Extensions)
+		}
+		for k, v := range rate.Ext {
+			c.Ext[k] = v
+		}
+	}
+
 	if rate.Exempt {
 		c.Percent = nil
 		c.Surcharge = nil
@@ -141,20 +153,17 @@ func (c *Combo) prepare(r *Regime, tags []cbc.Key, date cal.Date) error {
 		return ErrInvalidDate.WithMessage("rate value unavailable for '%s' in '%s' on '%s'", c.Rate.String(), c.Category.String(), date.String())
 	}
 
-	// 2024-03-14: only update the percentage if none previously set.
-	// This means that custom percentages can be used even if the
-	// rate classification is required by a regime (like PT).
-	if c.Percent == nil {
-		p := value.Percent // copy
-		c.Percent = &p
-	}
-	if c.Surcharge == nil {
-		if value.Surcharge != nil {
-			s := *value.Surcharge // copy
-			c.Surcharge = &s
-		} else {
-			c.Surcharge = nil
-		}
+	// Always overwrite the percentage. If a regime requires a rate
+	// to be set, this should be applied as an extension and not by
+	// forcing the rate key to be present.
+	p := value.Percent // copy
+	c.Percent = &p
+
+	if value.Surcharge != nil {
+		s := *value.Surcharge // copy
+		c.Surcharge = &s
+	} else {
+		c.Surcharge = nil
 	}
 
 	return nil